moved screen component
